requirements.txt
torch>=2.0.0
safetensors>=0.4.0
numpy
packaging
psutil
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
